+++ libraries/browser-tracker-core/src/tracker/types.ts
@@ -1,33 +1,3 @@
-/*
- * Copyright (c) 2022 Snowplow Analytics Ltd, 2010 Anthon Pang
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its
- *    contributors may be used to endorse or promote products derived from
- *    this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
 import { BrowserPlugin } from '../plugins';
 import {
   CommonEventProperties,
@@ -61,6 +31,20 @@ export type Platform = 'web' | 'mob' | 'pc' | 'srv' | 'app' | 'tv' | 'cnsl' | 'i
 export type CookieSameSite = 'None' | 'Lax' | 'Strict';
 /* The supported methods which events can be sent with */
 export type EventMethod = 'post' | 'get' | 'beacon';
+
+/* Available configuration for the extended cross domain linker */
+export type ExtendedCrossDomainLinkerAttributes = {
+  userId?: boolean;
+  sessionId?: boolean;
+  sourceId?: boolean;
+  sourcePlatform?: boolean;
+  /**
+   * Allow for the collection of the link text when a cross-domain link is clicked. Can also accept a callback for customizing information collection.
+   */
+  reason?: boolean | ((evt: Event) => string);
+};
+
+export type ExtendedCrossDomainLinkerOptions = boolean | ExtendedCrossDomainLinkerAttributes;
 
 /**
  * The configuration object for initialising the tracker
@@ -157,6 +141,11 @@ export type TrackerConfiguration = {
    */
   crossDomainLinker?: (elt: HTMLAnchorElement | HTMLAreaElement) => boolean;
   /**
+   * Configure the cross domain linker to use the extended format, allowing for
+   * more user/session information to pass to the cross domain navigation.
+   */
+  useExtendedCrossDomainLinker?: ExtendedCrossDomainLinkerOptions;
+  /**
    * The max size a POST request can be before the tracker will force send it
    * @defaultValue 40000
    */
@@ -249,6 +238,34 @@ export type TrackerConfiguration = {
    * The request respects the `anonymousTracking` option, including the SP-Anonymous header if needed, and any additional custom headers from the customHeaders option.
    */
   idService?: string;
+  /**
+   * Whether to retry failed requests to the collector.
+   *
+   * Failed requests are requests that failed due to
+   * [timeouts](https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest/timeout_event),
+   * [network errors](https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest/error_event),
+   * and [abort events](https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest/abort_event).
+   *
+   * Takes precedent over `retryStatusCodes` and `dontRetryStatusCodes`.
+   *
+   * @defaultValue true
+   */
+  retryFailedRequests?: boolean;
+  /**
+   * a callback function to be executed whenever a request is successfully sent to the collector.
+   * In practice this means any request which returns a 2xx status code will trigger this callback.
+   *
+   * @param data - The event batch that was successfully sent
+   */
+  onRequestSuccess?: (data: EventBatch) => void;
+
+  /**
+   * a callback function to be executed whenever a request fails to be sent to the collector.
+   * This is the inverse of the onRequestSuccess callback, so any non 2xx status code will trigger this callback.
+   *
+   * @param data - The data associated with the event(s) that failed to send
+   */
+  onRequestFailure?: (data: RequestFailure) => void;
 };
 
 /**
@@ -648,3 +665,35 @@ export interface ClientSession extends Record<string, unknown> {
    */
   firstEventTimestamp: string | null;
 }
+
+/**
+ * A collection of GET events which are sent to the collector.
+ * This will be a collection of query strings.
+ */
+export type GetBatch = string[];
+
+/**
+ * A collection of POST events which are sent to the collector.
+ * This will be a collection of JSON objects.
+ */
+export type PostBatch = Record<string, unknown>[];
+
+/**
+ * A collection of events which are sent to the collector.
+ * This can either be a collection of query strings or JSON objects.
+ */
+export type EventBatch = GetBatch | PostBatch;
+
+/**
+ * The data that will be available to the `onRequestFailure` callback
+ */
+export type RequestFailure = {
+  /** The batch of events that failed to send */
+  events: EventBatch;
+  /** The status code of the failed request */
+  status?: number;
+  /** The error message of the failed request */
+  message?: string;
+  /** Whether the tracker will retry the request */
+  willRetry: boolean;
+};