+++ trackers/node-tracker/test/got_emitter.ts
@@ -1,33 +1,3 @@
-/*
- * Copyright (c) 2022 Snowplow Analytics Ltd
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its
- *    contributors may be used to endorse or promote products derived from
- *    this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
 import test from 'ava';
 import sinon from 'sinon';
 import nock from 'nock';
@@ -35,25 +5,61 @@ import { HttpMethod, HttpProtocol, gotEmitter } from '../src/index';
 
 const endpoint = 'd3rkrsqld9gmqf.cloudfront.net';
 
-nock(new RegExp('https*://' + endpoint))
-  .persist()
-  .filteringPath(() => '/')
-  .get('/')
-  .reply(200, (uri) => uri);
-
-nock(new RegExp('https*://' + endpoint))
-  .matchHeader('content-type', 'application/json; charset=utf-8')
-  .persist()
-  .filteringRequestBody(() => '*')
-  .post('/com.snowplowanalytics.snowplow/tp2', '*')
-  .reply(200, (_uri, body: Record<string, unknown>) => (body['data'] as Array<unknown>)[0]);
-
 test.before(() => {
   nock.disableNetConnect();
 });
 
-test.after(() => {
+test.beforeEach(() => {
+  nock(new RegExp('https*://' + endpoint))
+    .filteringPath(() => '/')
+    .get('/')
+    .reply(200, (uri) => uri);
+
+  nock(new RegExp('https*://' + endpoint))
+    .matchHeader('content-type', 'application/json; charset=utf-8')
+    .filteringRequestBody(() => '*')
+    .post('/com.snowplowanalytics.snowplow/tp2', '*')
+    .reply(200, (_uri: string, body: Record<string, unknown>) => (body['data'] as Array<unknown>)[0]);
+});
+
+test.afterEach(() => {
   nock.cleanAll();
+});
+
+test.serial('gotEmitter should allow anonymization headers', async (t) => {
+  nock.cleanAll();
+
+  nock(new RegExp('https*://' + endpoint), {
+    reqheaders: {
+      'SP-Anonymous': '*',
+    },
+  })
+    .filteringPath(() => '/')
+    .get('/')
+    .once()
+    .reply(200, (uri) => uri);
+
+  await new Promise((resolve, reject) => {
+    const e = gotEmitter(
+      endpoint,
+      HttpProtocol.HTTPS,
+      80,
+      HttpMethod.GET,
+      undefined,
+      undefined,
+      undefined,
+      function (error, response) {
+        nock.cleanAll();
+        t.is(error, undefined);
+        t.pass();
+        if (error) reject(error);
+        else resolve(response);
+      },
+      undefined,
+      true
+    );
+    e.input({});
+  });
 });
 
 test('gotEmitter should send an HTTP GET request', async (t) => {