+++ plugins/browser-plugin-client-hints/CHANGELOG.md
@@ -1,6 +1,33 @@
 # Change Log - @snowplow/browser-plugin-client-hints
 
-This log was last generated on Mon, 16 Oct 2023 14:58:08 GMT and should not be manually modified.
+This log was last generated on Mon, 29 Jan 2024 08:34:06 GMT and should not be manually modified.
+
+## 3.21.0
+Mon, 29 Jan 2024 08:34:06 GMT
+
+_Version update only_
+
+## 3.20.0
+Mon, 15 Jan 2024 14:41:16 GMT
+
+_Version update only_
+
+## 3.19.0
+Thu, 14 Dec 2023 10:45:22 GMT
+
+_Version update only_
+
+## 3.18.0
+Mon, 04 Dec 2023 13:44:02 GMT
+
+_Version update only_
+
+## 3.17.0
+Tue, 14 Nov 2023 17:58:26 GMT
+
+### Updates
+
+- Drop invalid client hints
 
 ## 3.16.0
 Mon, 16 Oct 2023 14:58:08 GMT