+++ trackers/javascript-tracker/CHANGELOG.md
@@ -1,6 +1,37 @@
 # Change Log - @snowplow/javascript-tracker
 
-This log was last generated on Mon, 16 Oct 2023 14:58:08 GMT and should not be manually modified.
+This log was last generated on Mon, 29 Jan 2024 08:34:06 GMT and should not be manually modified.
+
+## 3.21.0
+Mon, 29 Jan 2024 08:34:06 GMT
+
+_Version update only_
+
+## 3.20.0
+Mon, 15 Jan 2024 14:41:16 GMT
+
+### Updates
+
+- Support clicks on child elements in button click tracking (#1280)
+
+## 3.19.0
+Thu, 14 Dec 2023 10:45:23 GMT
+
+_Version update only_
+
+## 3.18.0
+Mon, 04 Dec 2023 13:44:02 GMT
+
+### Updates
+
+- Add onRequestSuccess and onRequestFailure callbacks
+
+## 3.17.0
+Tue, 14 Nov 2023 17:58:26 GMT
+
+### Updates
+
+- Add further options for the gaCookies context
 
 ## 3.16.0
 Mon, 16 Oct 2023 14:58:08 GMT